requirements.txt
# =============================================================================
# Tech Watch Agent - Dépendances Python
# =============================================================================

# LangChain & LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0

# LLM Providers (APIs gratuites)
langchain-google-genai>=2.0.0  # Gemini
langchain-mistralai>=0.2.0     # Mistral

# HTTP & Async
httpx>=0.27.0
aiohttp>=3.9.0

# Parsing
beautifulsoup4>=4.12.0
feedparser>=6.0.0
lxml>=5.0.0

# Configuration
python-dotenv>=1.0.0

# Utilitaires
pydantic>=2.0.0

# =============================================================================
# Dépendances optionnelles (décommenter si besoin)
# =============================================================================

# Cache Redis (pour réduire les appels)
# redis>=5.0.0

# Monitoring
# prometheus-client>=0.20.0

# Tests
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
